requirements.txt
Django>=3.1
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2
channels
channels-redis
redis
django-redis
cryptography
pycryptodome
daphne
aioboto3
starlette